feat: move FileAperture class to polygon module and update imports
